+++ zoom/_assets/standard_apps/admin/index.py
@@ -83,7 +83,7 @@ def activity_panel(db):
         log.timestamp,
         log.elapsed
     from log left join users on log.user_id = users.id
-    where server = %s and path not like "%%\\/\\_%%"
+    where server = %s
     and log.status = 'C'
     order by timestamp desc
     limit 15
@@ -183,6 +183,7 @@ def users_panel(db):
     from users
     where
         last_seen >= %s
+        and username != 'guest'
     group by users.username
     order by last_seen desc
     limit 10